Fetch login & logout

+++ src/app/my/layout.jsx
@@ -1,25 +1,57 @@
 "use client";
 
-import { useRouter } from "next/navigation";
+import { usePathname, useRouter } from "next/navigation";
 import { useEffect, useState } from "react";
+import { getRole } from "../utils/getRole";
 
 export default function MyLayout({ children }) {
   const router = useRouter();
+  const pathname = usePathname();
 
   const [loading, setLoading] = useState(true);
 
-  const handleLogout = () => {
-    localStorage.clear();
-    router.push("/auth/login");
-  };
+  useEffect(() => {
+    const token = localStorage.getItem("token");
+    let role = localStorage.getItem("role");
 
-  // useEffect(() => {
-  //   const token = localStorage.getItem("token");
-  //   if (!token) {
-  //     router.push("/auth/login");
-  //     return;
-  //   }
-  // }, [router]);
+    if (!token) {
+      if (pathname != "/my/article" && pathname != "/my/library") {
+        router.push("/auth/login");
+        return;
+      }
+    }
 
-  return <>{children}</>;
+    if (!role && pathname != "/my/article" && pathname != "/my/library" && token) {
+      role = getRole();
+    }
+
+    if (role && token) {
+      if (role != "2" && role != "3") {
+        router.push("/dashboard");
+        return;
+      }
+    }
+
+    setLoading(false);
+  }, [router, pathname]);
+
+  useEffect(() => {
+    window.addEventListener("storage", async (event) => {
+      if (event.key === "token" || event.key === "role") {
+        const token = localStorage.getItem("token");
+        if (token) {
+          const role = await getRole();
+          if (role) {
+            localStorage.setItem("role", role);
+          }
+        } else {
+          localStorage.clear();
+        }
+
+        window.location.reload();
+      }
+    });
+  });
+
+  return <>{loading ? "Loading..." : children}</>;
 }
